requirements.txt
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
dnspython==2.7.0
executing==2.2.0
icecream==2.1.4
idna==3.10
paho-mqtt==2.1.0
passlib==1.7.4
pyaes==1.6.1
Pygments==2.19.1
pymongo==4.10.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
TgCrypto==1.2.5
urllib3==2.3.0
watchdog==6.0.0
